requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
boto==2.49.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
crcmod==1.7
cryptography==41.0.5
Django==5.0.1
djangorestframework==3.14.0
ecdsa==0.18.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
monotonic==1.6
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
retry_decorator==1.1.1
rsa==4.9
setuptools==70.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.4
uvicorn==0.24.0.post1
websockets==12.0
